Update MethodGraphFactory to take an `entriesToAllocate` parameter.
I *think* we could have calculated this from methodDataFlowGraph, but
time is short. I'm adding this parameter so we can use already-built
functionality without doing too much work.

Tests are now passing again.

+++ src/edu/mit/compilers/codegen/asm/instructions/Instructions.java
@@ -133,7 +133,7 @@ public final class Instructions {
      * Does `enter $x, $0`. Calculated appropriately for a method with the specified number of
      * entries.
      */
-    public static Instruction enter(int entries){
+    public static Instruction enter(long entries){
         return new Enter(entries);
     }
